\name{netvl-package}
\alias{netvl-package}
\alias{netvl}
\docType{package}
\title{
\packageTitle{netvl}
}
\description{
\packageDescription{netvl}
}
\details{
\packageDESCRIPTION{netvl}
\packageIndices{netvl}

All functions and methods of this package start with "netvl." and the naming convention used
throughtout the package is all lowercase period-separated. For instance, "netvl.plot.ts".
}
\author{
\packageAuthor{netvl}

Maintainer: \packageMaintainer{netvl}
}
\keyword{package}
